Remove the 1 GB limit on tile size.  This causes the issue432 unit test to succeed
when it really should not: when trying to view issue432.jp2,  kdu_show reports "illegal inclusion tag tree encountered while decoding a packet header", so we need to further investigate issue 432.

+++ src/lib/openjp2/tile_buf.h
@@ -45,9 +45,9 @@ typedef struct opj_tile_buf_resolution {
 typedef struct opj_tile_buf_component {
 	std::vector<opj_tile_buf_resolution_t*> resolutions;
     int32_t *data;
-    uint32_t data_size_needed;	/* we may either need to allocate this amount of data,
+    uint64_t data_size_needed;	/* we may either need to allocate this amount of data,
 									   or re-use image data and ignore this value */
-    uint32_t data_size;			/* size of the data of the component */
+    uint64_t data_size;			/* size of the data of the component */
     bool owns_data;				/* true if tile buffer manages its data array, false otherwise */
 
     opj_rect_t dim;		  /* canvas coordinates of region */
